+++ packages/jest-jasmine2/src/jasmine/JsApiReporter.js
