+++ packages/jest-jasmine2/src/jasmine/env.js
