+++ packages/jest-jasmine2/src/jasmine/spyStrategy.js
